test: move bitswap tests to a single file
